chore(article): implement search and filter
- add validator to validate and sanitize search parameters
-update GET /articles route with search validator middleware
- add helper function to format search query parameter
-update articlesController class to filter result based on search query parameters

[Finishes #164198195]
