+++ package.json
@@ -23,7 +23,7 @@
     "sequelize": "sequelize",
     "start": "node dist/bin/www.js",
     "start:dev": "nodemon --exec npm run dev",
-    "test": "nyc --reporter=html --reporter=text --reporter=lcov mocha --timeout 20000"
+    "test": "nyc --reporter=html --reporter=text --reporter=lcov mocha"
   },
   "author": "Andela Simulations Programme",
   "license": "MIT",
