+++ test/mocha.opts
@@ -2,5 +2,5 @@
 -r @babel/register
 -r @babel/polyfill
 -r test/setup
---timeout 50000
+--timeout 20000
 --exit
